Move ancestor_location expression tests to core
